Adding file "source\ScriptModules\ScriptModulesSourceFileIndex.h"
Reason, for easier overview.
