+++ AutoHotkeyx.vcxproj
@@ -265,6 +265,7 @@
     <ClInclude Include="source\ScriptModules\ScriptModulesDefines.h" />
     <ClInclude Include="source\ScriptModules\ModuleListMethods.h" />
     <ClInclude Include="source\ScriptModules\ScriptModulesReleaseObjects.h" />
+    <ClInclude Include="source\ScriptModules\ScriptModulesSourceFileIndex.h" />
     <ClInclude Include="source\script_com.h" />
     <ClInclude Include="source\script_func_impl.h" />
     <ClInclude Include="source\script_object.h" />
