+++ AutoHotkeyx.vcxproj.filters
@@ -219,6 +219,9 @@
     <ClInclude Include="source\ScriptModules\ScriptModulesReleaseObjects.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="source\ScriptModules\ScriptModulesSourceFileIndex.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="source\resources\icon_filetype.ico">
